requirements.txt
flask
flask-cors
pandas
python-dotenv
pdfplumber
markdownify
pyMuPDF
requests
flask_socketio
spacy